Lazy Pytorch Unpickler (#13)
Adds a flag `--lazy-unpickle` to `mergekit-yaml` that lazily loads
individual tensors instead of whole shards.

With this flag enabled I'm able to do a linear merge of two 13b models
in 16GB of RAM.

It's a little hacky and has to touch pytorch internals - I can't
guarantee it'll work right for 100% of models. Do let me know if it
breaks anything.