Generic Sparsification Rescaling (#245)
According to the [DARE paper](https://arxiv.org/pdf/2311.03099.pdf),
rescaling allows for the sparsified tensor to approximate the original
tensor. In the paper, this statement is made:
> We have also tried to combine MP with the rescale operation but got
worse results than using MP separately. This is because MP removes
parameters with smaller magnitudes and retains certain parameters with
the largest magnitudes. Simply rescaling the remaining ones would
destroy the original embeddings and result in unpredictable performance.
They claim that rescaling MP results in worse performance, but according
to their theory, it should work. I believe that that they may have used
the same rescaling operation as in the DARE method. However, the simple
rescaling method, 1/density, used for DARE only works due to dropping
making the sum of the weights equal to the sum times the density. For
sparsification methods that do not use random dropping, the necessary
rescaling changes based on the distribution of value in the weights.

In is due to this that I have implemented a generic rescaling method.
This also cleans up some of the code (No more rescaled random). To keep
compatibility, TIES still has no rescaling by default, but by adding
`rescale: true` to the parameters, you can enable it. Below is a test
using it:

<table>
<tr>
 <td>Model
 <td>MMLU
 <td>ARC
 <td>Winogrande
<tr>
 <td>TIES (rescaled)
 <td>51.08%
 <td>50.85%
 <td>75.53%
<tr>
 <td>DARE TIES
 <td>48.08%
 <td>45.39%
 <td>76.56%
</table>

This is not meant as proof that this method works. I am making this
change mostly to help make future development of sparsification methods
easier, since this rescaling algorithm works regardless of the
sparsification method. (Technically DARE could even be switched to use
it, but finding the necessary multiplier is more computationally
expensive than the simpler implementation that is only available with
dropping.)