Tokenizer merging overhaul (#334)
Rewrite the tokenizer merging logic to support all merge methods and
allow more customization of behavior.

The previous implementation of tokenizer merging always used either
linear or slerp to combine the embedding/LM head parameters. This was to
avoid the complexity that would be required to make all merge methods
support tensors that potentially have invalid or masked out values. It
works okay for some cases but wasn't a general solution.

In this implementation, instead of overriding the merge method for
embed/lm_head a preprocessing step remaps them to the vocabulary used by
the output model. These (now appropriately sized and ordered) tensors
are then merged normally.

The selection of embedding values for tokens not normally present in a
model is where things get slightly tricky. By default a set of
heuristics that I think are sane are applied. For a given token and
model, if the token is not present in the model's original tokenizer:
* If the base model has this token present, the base model's embedding
is used
* If only one model in the merge has the token, that model's embedding
is used
* Otherwise, the average of all embeddings for the token is assumed as a
default value

This can also be overridden on a per-token level. For example:

```yaml
merge_method: dare_ties
base_model: ...
models:
  - model: some_chatml_model
  - model: some_weird_model
  - model: some_model
tokenizer:
  source: union
  tokens:
    # if model doesn't have <|im_start|>, use embedding from some_chatml_model
    <|im_start|>:
      source: some_chatml_model
    # use embedding of <|special|> from some_weird_model for *all* models
    <|special|>:
      source: some_weird_model
      force: true
    # output tokenizer will have <|renamed_token|> with embedding of <|original_token|>
    # from some_model
    <|renamed_token|>:
      source:
        kind: model_token
        model: some_model
        token: <|original_token|>
      force: true
```

A practical example would be for merging two Llama 3 models, one using
the Llama 3 Instruct prompt format and one using chatml, trying to
preserve the ability to use both formats:
```yaml
tokenizer:
  source: union
  tokens:
    <|im_start|>:
      source: chatml_model
    <|im_end|>:
      source: chatml_model
    <|start_header_id|>:
      source: llama3_model
      force: true
    <|end_header_id|>:
      source: llama3_model
      force: true
    <|eot_id|>:
      source: llama3_model
      force: true
```